fix: avoids superflous WriteHeader.
