--- http/http.go
@@ -1,67 +1,0 @@
-// Copyright 2022 Juan Pablo Tosso and the OWASP Coraza contributors
-// SPDX-License-Identifier: Apache-2.0
-
-// tinygo does not support net.http so this package is not needed for it
-//go:build !tinygo
-// +build !tinygo
-
-package http
-
-import (
-	"io"
-	"net/http"
-	"strconv"
-	"strings"
-
-	"github.com/corazawaf/coraza/v3/types"
-)
-
-// processRequest fills all transaction variables from an http.Request object
-// Most implementations of Coraza will probably use http.Request objects
-// so this will implement all phase 0, 1 and 2 variables
-// Note: This function will stop after an interruption
-// Note: Do not manually fill any request variables
-func processRequest(tx types.Transaction, req *http.Request) (*types.Interruption, error) {
-	var (
-		client string
-		cport  int
-	)
-	// IMPORTANT: Some http.Request.RemoteAddr implementations will not contain port or contain IPV6: [2001:db8::1]:8080
-	idx := strings.LastIndexByte(req.RemoteAddr, ':')
-	if idx != -1 {
-		client = req.RemoteAddr[:idx]
-		cport, _ = strconv.Atoi(req.RemoteAddr[idx+1:])
-	}
-
-	var in *types.Interruption
-	// There is no socket access in the request object, so we neither know the server client nor port.
-	tx.ProcessConnection(client, cport, "", 0)
-	tx.ProcessURI(req.URL.String(), req.Method, req.Proto)
-	for k, vr := range req.Header {
-		for _, v := range vr {
-			tx.AddRequestHeader(k, v)
-		}
-	}
-	// Host will always be removed from req.Headers(), so we manually add it
-	if req.Host != "" {
-		tx.AddRequestHeader("Host", req.Host)
-	}
-
-	in = tx.ProcessRequestHeaders()
-	if in != nil {
-		return in, nil
-	}
-	if req.Body != nil {
-		_, err := io.Copy(tx.RequestBodyWriter(), req.Body)
-		if err != nil {
-			return tx.GetInterruption(), err
-		}
-		reader, err := tx.RequestBodyReader()
-		if err != nil {
-			return tx.GetInterruption(), err
-		}
-		req.Body = io.NopCloser(reader)
-	}
-
-	return tx.ProcessRequestBody()
-}
